Fix game not working on macOS with user account names lacking spaces
spaceIndex got assigned a -1 int, but NSNotFound is the max value of NSUInteger which breaks on x86_64

+++ src/osx.m
@@ -48,7 +48,7 @@ void getDefaultUserName(char *defaultUserName, int maxLength)
 	NSString *fullUsername = NSFullUserName();
 	if (fullUsername)
 	{
-		int spaceIndex = [fullUsername rangeOfString:@" "].location;
+		NSUInteger spaceIndex = [fullUsername rangeOfString:@" "].location;
 		if (spaceIndex != NSNotFound)
 		{
 			strncpy(defaultUserName, [[fullUsername substringToIndex:spaceIndex] UTF8String], maxLength);
